% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-size.r
\name{adjust.by.cell.sizes}
\alias{adjust.by.cell.sizes}
\title{Estimate the cell sizes and adjust the expression by cell size.}
\usage{
adjust.by.cell.sizes(dl)
}
\arguments{
\item{dl}{de.lorean object.}
}
\description{
Estimate the cell sizes and adjust the expression by cell size.
}

