% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{alpha.for.rug}
\alias{alpha.for.rug}
\title{Calculate a suitable value for a rug plot given the
number of points}
\usage{
alpha.for.rug(n, scale = 100)
}
\arguments{
\item{n}{Number of points.}

\item{scale}{Scale the value.}
}
\description{
Calculate a suitable value for a rug plot given the
number of points
}
