% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{analyse.noise.levels}
\alias{analyse.noise.levels}
\title{Analyse noise levels and assess which genes have the greatest
ratio of temporal variance to noise. This are labelled as the
'gene.high.psi' genes.}
\usage{
analyse.noise.levels(dl, num.high.psi = 25)
}
\arguments{
\item{dl}{de.lorean object}

\item{num.high.psi}{How many genes with high variance to examine}
}
\description{
Analyse noise levels and assess which genes have the greatest
ratio of temporal variance to noise. This are labelled as the
'gene.high.psi' genes.
}
