% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{analyse.variance}
\alias{analyse.variance}
\title{Analyse variance of expression between and within capture times.}
\usage{
analyse.variance(dl, adjust.cell.sizes)
}
\arguments{
\item{dl}{de.lorean object}

\item{adjust.cell.sizes}{Choose whether to adjust the expression values by
the cell size estimates}
}
\description{
Analyse variance of expression between and within capture times.
}
