% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-size.r
\name{anders.huber.cell.sizes}
\alias{anders.huber.cell.sizes}
\title{Estimate the cell sizes according to Anders & Huber
Differential expression analysis for sequence count data}
\usage{
anders.huber.cell.sizes(expr.l)
}
\arguments{
\item{expr.l}{Melted expression values.}
}
\description{
Estimate the cell sizes according to Anders & Huber
Differential expression analysis for sequence count data
}
