% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.r
\name{aov.dl}
\alias{aov.dl}
\title{Perform an analysis of variance to select genes for the DeLorean model.}
\usage{
aov.dl(dl, expr.l = melt.expr(dl))
}
\arguments{
\item{dl}{DeLorean object.}

\item{expr.l}{Melted (with melt.expr()) expression values to use}
}
\description{
Perform an analysis of variance to select genes for the DeLorean model.
}
