% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_add_ts.R
\name{archive_add_ts}
\alias{archive_add_ts}
\title{Add One or More Time Series to a Version}
\usage{
archive_add_ts(
  x,
  existing_version = NULL,
  version = Sys.Date(),
  root_folder = "../ts_archive",
  seal = FALSE
)
}
\arguments{
\item{x}{a single R time series object. Multiple time series objects can be represented as long format data.table or tslist.}

\item{version}{Date}

\item{root_folder}{character path to the time series archive. If the archive folder repository does not exist, it will be created.}
}
\description{
Add One or More Time Series to a Version
}
