% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_import_history.R
\name{archive_import_history}
\alias{archive_import_history}
\title{Import Full History of a Dataset from History Data.Table}
\usage{
archive_import_history(history_dt, repository_path)
}
\arguments{
\item{history_dt}{data.table with nested cells, contains key, release_date
and data in long format.}

\item{repository_path}{character path to local repository}
}
\description{
Import Full History of a Dataset from History Data.Table
}
