% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_init.R
\name{archive_init}
\alias{archive_init}
\title{Initialize a New Time Series Dataset Archive}
\usage{
archive_init(
  archive_name,
  parent_dir = NULL,
  remote_owner = "opentsi",
  rproj = TRUE,
  use_gha = TRUE
)
}
\description{
Initialize a New Time Series Dataset Archive
}
