% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_register_ts.R
\name{archive_register_ts}
\alias{archive_register_ts}
\title{Add New Time Series to an Existing Dataset (History Altering)}
\usage{
archive_register_ts(
  tsx,
  repo,
  repo_parent_dir = ".",
  owner = "opentsi",
  release_date_time = NULL,
  remote_provider = "https://github.com",
  clean_up_local = TRUE,
  local_only = FALSE
)
}
\description{
You can add new time series and all their versions (vintages) to an already
existing dataset, but you need to be aware of a few things:
}
\details{
Adding a version of a time series which was created before the
current version is not as innocent as it looks. You are altering the
history of your dataset by pretending this early version and all later
versions of said time series always existed.
Hence this functions adds the new series to from its first appearance on
to every commit (dataset version) that follows.

Second, keep in my mind that within a dataset all time series need to have
the same release date. If the release differs from the rest of the dataset
rather start a new dataset using archive_init().
}
