% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_update.R
\name{archive_update}
\alias{archive_update}
\title{Add a New Time Series Version}
\usage{
archive_update(
  sig = git_signature("Open Time Series Initiative", "bannert@kof.ethz.ch", Sys.Date()),
  push = FALSE
)
}
\arguments{
\item{sig}{git signature author and timesetamp for this particlar version}
}
\description{
Add a New Time Series Version
}
