% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{avg.par.samples}
\alias{avg.par.samples}
\title{Average across a parameters samples.}
\usage{
avg.par.samples(s)
}
\arguments{
\item{s}{An array of any dimension in which the first dimensions
indexes the samples}
}
\description{
Average across a parameters samples.
}
