% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.r
\name{branching.calc.log.marg.lik}
\alias{branching.calc.log.marg.lik}
\title{Calculate the log marginal likelihoods of each cell's expression
of each gene.}
\usage{
branching.calc.log.marg.lik(dl)
}
\arguments{
\item{dl}{The DeLorean object.}
}
\description{
Calculate the log marginal likelihoods of each cell's expression
of each gene.
}

