% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.r
\name{branching.gene.post}
\alias{branching.gene.post}
\title{Calculate the posterior for one gene in the branching model.}
\usage{
branching.gene.post(dl, g, psi, omega)
}
\arguments{
\item{dl}{The DeLorean object}

\item{g}{The index of the gene in the expression matrix (gene.map)}

\item{psi}{The temporal variation of the gene}

\item{omega}{The noise level of the gene}
}
\description{
Calculate the posterior for one gene in the branching model.
}

