% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.r
\name{branching.genes.post}
\alias{branching.genes.post}
\title{Calculate the posterior for several genes in the branching model.}
\usage{
branching.genes.post(dl, genes)
}
\arguments{
\item{dl}{The DeLorean object}

\item{genes}{The genes to calculate the posterior for}
}
\description{
Calculate the posterior for several genes in the branching model.
}

