% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{branching.post.plot}
\alias{branching.post.plot}
\title{Plot the posterior of the branching model}
\usage{
branching.post.plot(dl, post, ...)
}
\arguments{
\item{dl}{The DeLorean object}

\item{post}{The posterior (output of branching.genes.post)}

\item{...}{Arguments to be passed to geom_point(aes(...))}
}
\description{
Plot the posterior of the branching model
}

