% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.r
\name{branching.prepare.posterior}
\alias{branching.prepare.posterior}
\title{Prepare to calculate posteriors across the latent space in the branching model.}
\usage{
branching.prepare.posterior(dl, sample.iter = dl$best.sample,
  tau.grid.size = 51, z.grid.size = 51)
}
\arguments{
\item{dl}{The DeLorean object}

\item{sample.iter}{The sample (iteration) to use}

\item{tau.grid.size}{The number of grid entries to use in tau dimension}

\item{z.grid.size}{The number of grid entries to use in z dimension}
}
\description{
Prepare to calculate posteriors across the latent space in the branching model.
}

