% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{branching.tau.z.plot}
\alias{branching.tau.z.plot}
\title{Plot tau vs. z for the branching model.}
\usage{
branching.tau.z.plot(dl, sample.iter = dl$best.sample, ...)
}
\arguments{
\item{dl}{The DeLorean object}

\item{sample.iter}{The sample (iteration) to use}

\item{...}{Arguments to be passed to ggplot(aes(...))}
}
\description{
Plot tau vs. z for the branching model.
}

