% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{calc.cell.ll}
\alias{calc.cell.ll}
\title{Calculate the log likelihoods for each cell}
\usage{
calc.cell.ll(dl)
}
\arguments{
\item{dl}{The DeLorean object.}
}
\description{
Calculate the log likelihoods for each cell
}

