% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{calc.gene.ll}
\alias{calc.gene.ll}
\title{Calculate the log likelihoods for each gene}
\usage{
calc.gene.ll(dl)
}
\arguments{
\item{dl}{The DeLorean object.}
}
\description{
Calculate the log likelihoods for each gene
}
