% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{calc.log.marg.lik}
\alias{calc.log.marg.lik}
\title{Calculate the GP log marginal likelihoods for the data given the parameters
from the best sample. Also calculate the log likelihoods of these parameters.}
\usage{
calc.log.marg.lik(dl)
}
\arguments{
\item{dl}{The DeLorean object.}
}
\description{
Calculate the GP log marginal likelihoods for the data given the parameters
from the best sample. Also calculate the log likelihoods of these parameters.
}
