% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothness.r
\name{calc.roughness}
\alias{calc.roughness}
\title{Calculate the roughness of the vector. The roughness is the RMS
of the differences between consecutive points.}
\usage{
calc.roughness(x)
}
\arguments{
\item{x}{Values}
}
\description{
Calculate the roughness of the vector. The roughness is the RMS
of the differences between consecutive points.
}
