% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{centralise}
\alias{centralise}
\title{Centralises a periodic position into [period/2, period)
by shifting by n*period, where n is an integer}
\usage{
centralise(x, period = 1)
}
\arguments{
\item{x}{Position}

\item{period}{Period}
}
\description{
Centralises a periodic position into [period/2, period)
by shifting by n*period, where n is an integer
}

