% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{cmp.profiles.plot}
\alias{cmp.profiles.plot}
\title{Plot a comparison of the profiles from several de.lorean objects}
\usage{
cmp.profiles.plot(..., genes = NULL)
}
\arguments{
\item{...}{Named de.lorean objects}

\item{genes}{Genes to plot (defaults to genes.high.psi of first de.lorean
object)}
}
\description{
Plot a comparison of the profiles from several de.lorean objects
}
