% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_common.R
\name{commit_by_date}
\alias{commit_by_date}
\title{Register Commit with Specified Date in the Commit Signature}
\usage{
commit_by_date(
  repo = NULL,
  date_time = Sys.time(),
  tz = "UTC",
  author = "Open Time Series Initiative",
  email = "bannert@kof.ethz.ch"
)
}
\arguments{
\item{repo}{character repository}

\item{date_time}{POSIXct of the time}

\item{tz}{character timezone. Defaults to UTC.}

\item{author}{character commit author, will appear in the commit signature.
Will use env vars when set to NULL.}

\item{email}{character commit author email, will appear in the commit signature
Will use env vars when set to NULL.}
}
\description{
This helper functions is typically not called directly. It helps
to organize versions of entire datasets by date.
}
