% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{compile.model}
\alias{compile.model}
\title{Compile the model and cache the DSO to avoid unnecessary recompilation.}
\usage{
compile.model(dl)
}
\arguments{
\item{dl}{de.lorean object}
}
\description{
Compile the model and cache the DSO to avoid unnecessary recompilation.
}

