% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{cov.calc.dists}
\alias{cov.calc.dists}
\title{Calculate distances between vectors of time points}
\usage{
cov.calc.dists(tau.1, tau.2 = tau.1, period = NULL)
}
\arguments{
\item{tau.1}{First vector of time points}

\item{tau.2}{Second vector of time points (defaults to first if not given)}

\item{period}{Period if periodic}
}
\description{
Calculate distances between vectors of time points
}

