% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{cov.calc.dl.dists}
\alias{cov.calc.dl.dists}
\title{Calculate distances over estimated pseudotimes and
test inputs.}
\usage{
cov.calc.dl.dists(dl, tau = tau.for.sample(dl), include.test = TRUE)
}
\arguments{
\item{dl}{de.lorean object}

\item{tau}{The pseudotimes to use}

\item{include.test}{Also include the pseudotimes for the test inputs}
}
\description{
Calculate distances over estimated pseudotimes and
test inputs.
}
