% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{cov.calc.gene}
\alias{cov.calc.gene}
\title{Calculate covariance structure for gene over pseudotimes and test
inputs.}
\usage{
cov.calc.gene(dl, gene.idx, cov.fn = cov.matern.32,
  tau = tau.for.sample(dl), include.test = TRUE,
  psi = sampled.gene.param(dl, gene.idx, "psi"),
  omega = sampled.gene.param(dl, gene.idx, "omega"))
}
\arguments{
\item{dl}{de.lorean object}

\item{gene.idx}{Gene index}

\item{cov.fn}{Covariance function (defaults to cov.matern.32)}

\item{tau}{The pseudotimes to use}

\item{include.test}{Also include the pseudotimes for the test inputs}

\item{psi}{Temporal variation}

\item{omega}{Noise}
}
\description{
Calculate covariance structure for gene over pseudotimes and test
inputs.
}
