% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{create.ordering.ll.fn}
\alias{create.ordering.ll.fn}
\title{Calculate the covariance structure of evenly spread tau and
create a function that calculates the log likelihood of
orderings.}
\usage{
create.ordering.ll.fn(dl, cov.fn = cov.fn.for(dl))
}
\arguments{
\item{dl}{The DeLorean object}

\item{cov.fn}{The covariance function}
}
\description{
Calculate the covariance structure of evenly spread tau and
create a function that calculates the log likelihood of
orderings.
}

