% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_common.R
\name{create_vintages}
\alias{create_vintages}
\title{Create Vintages From a Named List of Time Series}
\usage{
create_vintages(release_date, tsl, time_chunk = "23:59:59", tz = "UTC")
}
\description{
Create Vintages From a Named List of Time Series
}
