% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_update.R
\name{dataset_update}
\alias{dataset_update}
\title{Handle Regular Updates of Dataset}
\usage{
dataset_update(
  tsx,
  repo,
  repo_parent_dir = ".",
  owner = "opentsi",
  release_date_time = NULL,
  remote_provider = "https://github.com",
  clean_up_local = TRUE,
  local_only = FALSE
)
}
\arguments{
\item{tsx}{object containing time series represented in either ts, tslist,
list of time series.}
}
\description{
opentsi archives use git to manage versioned time series. Once
a dataset archive is initialized and the full history of dataset is imported,
datasets need to be updated on a regular basis. This dataset_update function
is designed to do just that. It is meant to run in a batch process and
append to the next iteration of the time series in a dataset to the history.
}
