% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{de.lorean}
\alias{de.lorean}
\title{Initialise DeLorean object}
\usage{
de.lorean(expr, gene.meta, cell.meta)
}
\arguments{
\item{expr}{Expression array}

\item{gene.meta}{Data frame of meta data for genes}

\item{cell.meta}{Data frame of meta data for cells}
}
\description{
Initialise DeLorean object
}
\examples{
data(WindramDeLorean)
dl <- de.lorean(windram.expr, windram.gene.meta, windram.cell.meta)

}
