% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-size.r
\name{estimate.capture.cell.sizes}
\alias{estimate.capture.cell.sizes}
\title{Estimate the cell sizes per capture. Only uses genes that are expressed
in more than half the cells.}
\usage{
estimate.capture.cell.sizes(expr.l)
}
\arguments{
\item{expr.l}{Melted expression values.}
}
\description{
Estimate the cell sizes per capture. Only uses genes that are expressed
in more than half the cells.
}

