% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{examine.convergence}
\alias{examine.convergence}
\title{Analyse the samples and gather the convergence statistics. Note this
only makes sense if a sampling method was used to fit the model as
opposed to variational Bayes.}
\usage{
examine.convergence(dl)
}
\arguments{
\item{dl}{de.lorean object}
}
\description{
Analyse the samples and gather the convergence statistics. Note this
only makes sense if a sampling method was used to fit the model as
opposed to variational Bayes.
}
