% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{expected.sample.var}
\alias{expected.sample.var}
\title{The expected within sample variance of a Gaussian with the given covariance.}
\usage{
expected.sample.var(K)
}
\arguments{
\item{K}{Covariance}
}
\description{
The expected within sample variance of a Gaussian with the given covariance.
}
