% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{expr.data.plot}
\alias{expr.data.plot}
\title{Plot the expression data by the capture points}
\usage{
expr.data.plot(dl, genes = NULL, num.genes = 12)
}
\arguments{
\item{dl}{de.lorean object}

\item{genes}{Genes to plot. If NULL plots some random varying genes}

\item{num.genes}{Number of genes to plot}
}
\description{
Plot the expression data by the capture points
}
