% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{filter.cells}
\alias{filter.cells}
\title{Filter cells}
\usage{
filter.cells(dl, .filter = function(x) x \%in\% cells, number = NULL,
  cells = sample(colnames(dl$expr), number))
}
\arguments{
\item{dl}{de.lorean object}

\item{.filter}{Function that gakes a list of cells as input and returns
a vector of TRUE/FALSE}

\item{number}{Number to sample if filter function or cells not supplied.}

\item{cells}{The cells to keep.}
}
\description{
Filter cells
}

