% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{filter_genes}
\alias{filter_genes}
\title{Filter genes}
\usage{
filter_genes(dl, .filter = function(x) x \%in\% genes, number = NULL,
  genes = sample(rownames(dl$expr), number))
}
\arguments{
\item{dl}{de.lorean object}

\item{.filter}{Function that gakes a list of genes as input and returns
a vector of TRUE/FALSE}

\item{number}{Number to sample if filter function or genes not supplied.}

\item{genes}{The genes to keep.}
}
\description{
Filter genes
}
\examples{
data(WindramDeLorean)
dl <- de.lorean(windram.expr, windram.gene.meta, windram.cell.meta)
dl <- filter_genes(dl, number = 37)

}
