% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{find.best.tau}
\alias{find.best.tau}
\title{Find best tau to initialise chains with by sampling tau from the prior
and using empirical Bayes parameter estimates for the other parameters.}
\usage{
find.best.tau(dl, num.tau.candidates = 6000,
  num.tau.to.keep = num.cores, num.cores = default.num.cores())
}
\arguments{
\item{dl}{de.lorean object}

\item{num.tau.candidates}{How many candidates to examine. Defaults to 6000.}

\item{num.tau.to.keep}{How many candidates to keep. Defaults to num.cores.}

\item{num.cores}{Number of cores to run on. Defaults to default.num.cores()}
}
\description{
Find best tau to initialise chains with by sampling tau from the prior
and using empirical Bayes parameter estimates for the other parameters.
}
