% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{find.good.ordering}
\alias{find.good.ordering}
\title{Run a find good ordering method and append results to existing orderings}
\usage{
find.good.ordering(dl, method, ...)
}
\arguments{
\item{dl}{de.lorean object}

\item{method}{Function that runs the method}

\item{...}{Any other arguments for the method}
}
\description{
Run a find good ordering method and append results to existing orderings
}
