% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{fit.held.out}
\alias{fit.held.out}
\title{Fit held out genes}
\usage{
fit.held.out(dl, expr.held.out, sample.iter = dl$best.sample)
}
\arguments{
\item{dl}{de.lorean object}

\item{expr.held.out}{The expression matrix including the held out genes}

\item{sample.iter}{The sample to use to fit with

Calculate covariance over pseudotimes and capture times
Evaluate the held out gene under the GP model using pseudotimes
and a model without.}
}
\description{
Fit held out genes
}
