% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{fit.model}
\alias{fit.model}
\title{Fit the model using specified method (sampling or variational Bayes).}
\usage{
fit.model(dl, method = "sample", ...)
}
\arguments{
\item{dl}{de.lorean object}

\item{method}{Fitting method:
\itemize{
  \item 'sample': Use a Stan sampler.
    See \code{\link{fit.model.sample}}.
  \item 'vb': Use Stan ADVI variational Bayes algorithm.
    See \code{\link{fit.model.vb}}.
}}

\item{...}{Extra arguments for method}
}
\description{
Fit the model using specified method (sampling or variational Bayes).
}
