% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{fit.model.sample}
\alias{fit.model.sample}
\title{Fit the model using Stan sampler}
\usage{
fit.model.sample(dl, num.cores = getOption("DL.num.cores",
  max(parallel::detectCores() - 1, 1)), chains = 1, thin = 50, ...)
}
\arguments{
\item{dl}{de.lorean object}

\item{num.cores}{Number of cores to run on.
Defaults to getOption("DL.num.cores", max(parallel::detectCores()-1, 1))}

\item{chains}{Number of chains to run on each core}

\item{thin}{How many samples to generate before retaining one}

\item{...}{Extra arguments for rstan::stan() sampling call}
}
\description{
Fit the model using Stan sampler
}

