% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{fit.model.vb}
\alias{fit.model.vb}
\title{Fit the model using Stan variational Bayes}
\usage{
fit.model.vb(dl, num.cores = default.num.cores(),
  num.inits = num.cores, init.idx = 1, ...)
}
\arguments{
\item{dl}{de.lorean object}

\item{num.cores}{Number of cores to run on. Defaults to default.num.cores()}

\item{num.inits}{Number initialisations to try. Defaults to num.cores}

\item{init.idx}{Which initialisation to use if only using one}

\item{...}{Extra arguments for rstan::vb()}
}
\description{
Fit the model using Stan variational Bayes
}
