% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{gaussian.condition}
\alias{gaussian.condition}
\title{Condition a Gaussian on another.
See Eqn. A.6
on page 200 of Rasmumssen and Williams' book.}
\usage{
gaussian.condition(y, .A, .B, .C, mu.x = rep(0, nrow(.A)), mu.y = rep(0,
  nrow(.B)), U = chol(.B))
}
\arguments{
\item{y}{y}

\item{.A}{Var(X)}

\item{.B}{Var(Y)}

\item{.C}{Cov(X, Y)}

\item{mu.x}{Mean of x}

\item{mu.y}{Mean of y}

\item{U}{Cholesky decomposition of .B}
}
\description{
Condition a Gaussian on another.
See Eqn. A.6
on page 200 of Rasmumssen and Williams' book.
}
