% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{gene.covariances}
\alias{gene.covariances}
\title{Calculate the covariance structure of the tau}
\usage{
gene.covariances(cov.fn, length.scale, period = NULL, tau, psi, omega)
}
\arguments{
\item{cov.fn}{Covariance function}

\item{length.scale}{Length scale}

\item{period}{The period of the covariance function}

\item{tau}{Pseudotimes}

\item{psi}{The temporal variation for each gene}

\item{omega}{The noise level for each gene}
}
\description{
Calculate the covariance structure of the tau
}

