% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{gene.params.plot}
\alias{gene.params.plot}
\title{Plot the posterior of the gene parameters.}
\usage{
gene.params.plot(dl, alpha = 0.4)
}
\arguments{
\item{dl}{de.lorean object}

\item{alpha}{Transparency for points and error bars}
}
\description{
Plot the posterior of the gene parameters.
}

