% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{get.posterior.mean}
\alias{get.posterior.mean}
\title{Get posterior mean of samples}
\usage{
get.posterior.mean(extract)
}
\arguments{
\item{extract}{A named list of samples}
}
\description{
Get posterior mean of samples
}
