% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_georef_helpers.R
\name{get_georef_helpers}
\alias{get_georef_helpers}
\title{Download BDTOPO helper layers for manual georeferencing}
\usage{
get_georef_helpers(x, ..., filename = NULL)
}
\arguments{
\item{x}{Object of class \code{sf} or \code{sfc}}

\item{...}{\strong{category names} to select which helper layers to download:
\itemize{
\item \code{"hydro"}: hydrography (nodes, river segments, lake, pond…)
\item \code{"road"}: roads and railway lines)
\item \code{"infra"}: linear / point / surface infrastructure
\item \code{"building"}: building footprints
\item \code{"other"}: miscellaneous landmarks (e.g. reference points)
}}

\item{filename}{\code{character} or \code{NULL}; optional path to a file. If supplied, all
downloaded layers are written to a single \strong{GeoPackage} at this location.
Extension is automatically forced to \code{.gpkg}.}
}
\value{
A \strong{named list} of \code{sf} objects.
}
\description{
Download BDTOPO helper layers for manual georeferencing
}
\examples{
\dontrun{
library(sf)
x <- read_sf(system.file("extdata/penmarch.shp", package = "happign"))
hydro <- get_georef_helpers(x, "hydro")
all <- get_georef_helpers(x)

gpkg <- tempfile(fileext = ".gpkg")
save_to_gpkg <- get_georef_helpers(x, "road", "hydro", filename = gpkg)
}

}
