% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_metadata.R
\name{get_iso_639_1_codes}
\alias{get_iso_639_1_codes}
\title{Get ISO 639-1 Language Codes}
\usage{
get_iso_639_1_codes()
}
\value{
Character vector of valid two-letter language codes.
}
\description{
Returns a character vector of valid ISO 639-1 two-letter language codes.
Requires the ISOcodes package.
}
\keyword{internal}
