% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_meta.R
\name{get_label}
\alias{get_label}
\title{Get Label from Multilingual Object}
\usage{
get_label(obj, lang = "en")
}
\arguments{
\item{obj}{Multilingual text object.}

\item{lang}{Language code.}
}
\value{
Character string.
}
\description{
Extracts label for specified language, falls back to English.
}
\keyword{internal}
