% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{gp.log.marg.like.individual}
\alias{gp.log.marg.like.individual}
\title{The log marginal likelihood for each individual data point.
See "2.3 Varying the Hyperparameters"
on page 19 of Rasmumssen and Williams' book.}
\usage{
gp.log.marg.like.individual(y, K = NULL, U = chol(K))
}
\arguments{
\item{y}{The targets.}

\item{K}{The covariance matrix (kernel), not needed if U is provided.}

\item{U}{Cholesky decomposition of K (chol(K)).}
}
\description{
The log marginal likelihood for each individual data point.
See "2.3 Varying the Hyperparameters"
on page 19 of Rasmumssen and Williams' book.
}
