% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{gp.log.marg.like}
\alias{gp.log.marg.like}
\title{The log marginal likelihood. See "2.3 Varying the Hyperparameters"
on page 19 of Rasmussen and Williams' book.}
\usage{
gp.log.marg.like(y, K = NULL, U = chol(K))
}
\arguments{
\item{y}{The targets.}

\item{K}{The covariance matrix (kernel), not needed if U is provided.}

\item{U}{Cholesky decomposition of K (chol(K)).}
}
\description{
The log marginal likelihood. See "2.3 Varying the Hyperparameters"
on page 19 of Rasmussen and Williams' book.
}
