% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{gp.predict}
\alias{gp.predict}
\title{Predictive mean, variance and log marginal likelihood of a GP.
See "2.3 Varying the Hyperparameters"
on page 19 of Rasmussen and Williams' book.}
\usage{
gp.predict(y, K = NULL, Kstar, Kstarstar, U = chol(K))
}
\arguments{
\item{y}{The targets.}

\item{K}{The covariance matrix (kernel) for input points, not needed if U is provided.}

\item{Kstar}{The cross covariance matrix (kernel)}

\item{Kstarstar}{The cross covariance matrix (kernel) for test points}

\item{U}{Cholesky decomposition of K}
}
\description{
Predictive mean, variance and log marginal likelihood of a GP.
See "2.3 Varying the Hyperparameters"
on page 19 of Rasmussen and Williams' book.
}
