% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{gp.predictions.df}
\alias{gp.predictions.df}
\title{Convert the output of gp.predict() into a data.frame.}
\usage{
gp.predictions.df(predictions)
}
\arguments{
\item{predictions}{The predictions}
}
\description{
Convert the output of gp.predict() into a data.frame.
}

