% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{guo.expr}
\alias{guo.cell.meta}
\alias{guo.expr}
\alias{guo.gene.meta}
\title{Single cell expression data and meta data from Guo et al. (2012).
They investigated the expression of 48 genes in 500 mouse embryonic cells.}
\format{There are three objects in this data:
\itemize{
  \item guo.expr A matrix of log expression values with
    no missing data. Rows are named by genes and columns are
    named by cells/samples.
  \item guo.gene.meta A data frame containing meta-data
    about the genes.
  \item guo.cell.meta A data frame containing meta-data
    about the cells
}}
\source{
\url{http://www.sciencedirect.com/science/article/pii/S1534580710001103}
}
\usage{
data(GuoDeLorean)
}
\description{
Single cell expression data and meta data from Guo et al. (2012).
They investigated the expression of 48 genes in 500 mouse embryonic cells.
}
\keyword{datasets}

