% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/held-out.r
\name{held.out.melt}
\alias{held.out.melt}
\title{Melt held out genes}
\usage{
held.out.melt(dl, expr, held.out.genes)
}
\arguments{
\item{dl}{de.lorean object}

\item{expr}{Expression matrix of all genes}

\item{held.out.genes}{Genes to hold out}
}
\description{
Melt held out genes
}

