% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/held-out.r
\name{held.out.posterior.by.variation}
\alias{held.out.posterior.by.variation}
\title{Order the genes by the variation of their posterior mean}
\usage{
held.out.posterior.by.variation(posterior)
}
\arguments{
\item{posterior}{The posterior of some held out genes}
}
\description{
Order the genes by the variation of their posterior mean
}
