% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/held-out.r
\name{held.out.posterior.filter}
\alias{held.out.posterior.filter}
\title{Filter the genes}
\usage{
held.out.posterior.filter(posterior, genes)
}
\arguments{
\item{posterior}{The posterior of some held out genes}

\item{genes}{Genes to filter}
}
\description{
Filter the genes
}
