% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/held-out.r
\name{held.out.posterior.join}
\alias{held.out.posterior.join}
\title{Join with another data frame. Useful for adding gene names etc..}
\usage{
held.out.posterior.join(posterior, .df)
}
\arguments{
\item{posterior}{The posterior of some held out genes}

\item{.df}{Data frame to join with}
}
\description{
Join with another data frame. Useful for adding gene names etc..
}
