% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/held-out.r
\name{held.out.select.genes}
\alias{held.out.select.genes}
\title{Select held out genes by those with highest variance}
\usage{
held.out.select.genes(dl, expr, num.held.out)
}
\arguments{
\item{dl}{de.lorean object}

\item{expr}{Expression matrix of all genes}

\item{num.held.out}{Number to select}
}
\description{
Select held out genes by those with highest variance
}

