% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{inducing.covariance}
\alias{inducing.covariance}
\title{Calculate the covariance structure of the inducing points}
\usage{
inducing.covariance(cov.fn, length.scale, period = NULL, tau, num.inducing,
  u = seq(min(tau), max(tau), length.out = num.inducing))
}
\arguments{
\item{cov.fn}{Covariance function}

\item{length.scale}{Length scale}

\item{period}{The period of the covariance function}

\item{tau}{Pseudotimes}

\item{num.inducing}{How many inducing points to use}

\item{u}{The inducing points}
}
\description{
Calculate the covariance structure of the inducing points
}

