% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{init.orderings.vs.pseudotimes.plot}
\alias{init.orderings.vs.pseudotimes.plot}
\title{Plot the orderings for initialisation against the estimated pseudotime.}
\usage{
init.orderings.vs.pseudotimes.plot(dl, sample.iter = dl$best.sample)
}
\arguments{
\item{dl}{The DeLorean object}

\item{sample.iter}{Which sample to take pseudotimes from}
}
\description{
Plot the orderings for initialisation against the estimated pseudotime.
}
