% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_meta_data_catalog.R
\name{init_meta_data_catalog}
\alias{init_meta_data_catalog}
\title{Initialize a Metadata Catalog}
\usage{
init_meta_data_catalog(
  variables,
  country,
  provider,
  dataset,
  frequency,
  url = NULL,
  mi_languages = c("en", "de", "fr", "it")
)
}
\arguments{
\item{variables}{character vector of variable names. Variable names have to be unique
within the same dataset.}
}
\description{
Initialize a Metadata Catalog
}
