% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data_processing.R
\name{is_update_needed}
\alias{is_update_needed}
\title{Checks Whether a Time Series Archive Is Up to Date}
\usage{
is_update_needed(publisher_last_update, data_dir = "data-raw")
}
\arguments{
\item{publisher_last_update}{POSIXct formatted UTC time of publisher's last
available publication}
}
\description{
This functions returns TRUE when a more recent version is available.
}
