% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{knit.report}
\alias{knit.report}
\title{Knit a report, the file inst/Rmd/<report.name>.Rmd must exist in
the package directory.}
\usage{
knit.report(dl, report.name)
}
\arguments{
\item{dl}{de.lorean object}

\item{report.name}{The name of the report. Used to locate the R
markdown report file in the package.}
}
\description{
Knit a report, the file inst/Rmd/<report.name>.Rmd must exist in
the package directory.
}
