% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{kouno.expr}
\alias{kouno.expr}
\alias{kouno.cell.meta}
\alias{kouno.gene.meta}
\title{Kouno et al. investigated the transcriptional network controlling how
THP-1 human myeloid monocytic leukemia cells differentiate into
macrophages. They provide expression values for 45 genes in 960 single
cells captured across 8 distinct time points.}
\format{There are three objects in this data:
\itemize{
  \item kouno.expr A matrix of log expression values with
    no missing data. Rows are named by genes and columns are
    named by cells/samples.
  \item kouno.gene.meta A data frame containing meta-data
    about the genes.
  \item kouno.cell.meta A data frame containing meta-data
    about the cells
}}
\source{
\url{http://genomebiology.com/2013/14/10/R118/abstract}
}
\usage{
data(KounoDeLorean)
}
\description{
Kouno et al. investigated the transcriptional network controlling how
THP-1 human myeloid monocytic leukemia cells differentiate into
macrophages. They provide expression values for 45 genes in 960 single
cells captured across 8 distinct time points.
}
\keyword{datasets}
