% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.r
\name{magda.find.orderings}
\alias{magda.find.orderings}
\title{Use Magda's code to find good orderings}
\usage{
magda.find.orderings(dl, number_paths = 5)
}
\arguments{
\item{dl}{de.lorean object}

\item{number_paths}{Number of paths for each starting point}
}
\description{
Use Magda's code to find good orderings
}

