% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{make.fit.valid}
\alias{make.fit.valid}
\title{Make a fit valid by running one iteration of the sampler.}
\usage{
make.fit.valid(dl)
}
\arguments{
\item{dl}{de.lorean object}
}
\description{
Make a fit valid by running one iteration of the sampler.
}
