% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.r
\name{make.init.fn}
\alias{make.init.fn}
\title{Returns a function that constructs parameter settings with good tau if they are available.}
\usage{
make.init.fn(dl)
}
\arguments{
\item{dl}{de.lorean object}
}
\description{
Returns a function that constructs parameter settings with good tau if they are available.
}

