% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{marg.like.plot}
\alias{marg.like.plot}
\title{Plot posterior for marginal log likelihoods of individual gene's
expression profiles}
\usage{
marg.like.plot(dl)
}
\arguments{
\item{dl}{de.lorean object}
}
\description{
Plot posterior for marginal log likelihoods of individual gene's
expression profiles
}
