% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{melt.expr}
\alias{melt.expr}
\title{Melt an expression matrix.}
\usage{
\method{melt}{expr}(dl, expr = dl$expr)
}
\arguments{
\item{dl}{The de.lorean object.}

\item{expr}{Matrix of expression values.}
}
\description{
Melt an expression matrix.
}
