% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{mutate.profile.data}
\alias{mutate.profile.data}
\title{Mutate the profile data into shape compatible with GP plot function}
\usage{
\method{mutate}{profile.data}(.data)
}
\arguments{
\item{.data}{The data}
}
\description{
Mutate the profile data into shape compatible with GP plot function
}
