% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{omega.post.plot}
\alias{omega.post.plot}
\title{Plot the posterior of the noise levels and show the prior}
\usage{
omega.post.plot(dl)
}
\arguments{
\item{dl}{The DeLorean object}
}
\description{
Plot the posterior of the noise levels and show the prior
}
