% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{optimise.best.sample}
\alias{optimise.best.sample}
\title{Optimise the best sample and update the best.sample index.}
\usage{
optimise.best.sample(dl, sample.to.opt = dl$best.sample,
  new.best.sample = -1)
}
\arguments{
\item{dl}{de.lorean object}

\item{sample.to.opt}{Sample to optimise}

\item{new.best.sample}{Update to best sample index}
}
\description{
Optimise the best sample and update the best.sample index.
}
