% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.r
\name{ordering.block.move}
\alias{ordering.block.move}
\title{Move a block in an ordering and shift the other items.}
\usage{
ordering.block.move(ordering, from, width, to, reverse = FALSE)
}
\arguments{
\item{ordering}{The ordering.}

\item{from}{The start index of the block to move.}

\item{width}{The width of the block to move.}

\item{to}{The index it should be moved to.}

\item{reverse}{Reverse the block?}
}
\description{
Move a block in an ordering and shift the other items.
}
