% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.r
\name{ordering.improve}
\alias{ordering.improve}
\title{Improve the ordering in the sense that some function is
maximised.}
\usage{
ordering.improve(fn, ordering)
}
\arguments{
\item{fn}{A function to maximise.}

\item{ordering}{The permutation (ordering) to start from.}
}
\description{
Improve the ordering in the sense that some function is
maximised.
}
