% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.r
\name{ordering.invert}
\alias{ordering.invert}
\title{Invert the ordering}
\usage{
ordering.invert(ordering)
}
\arguments{
\item{ordering}{The permutation (ordering) to invert.}
}
\description{
Invert the ordering
}
