% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.r
\name{ordering.is.valid}
\alias{ordering.is.valid}
\title{Check that it is a valid ordering}
\usage{
ordering.is.valid(ordering, full.check = FALSE)
}
\arguments{
\item{ordering}{The ordering}

\item{full.check}{Perform a full check}
}
\description{
Check that it is a valid ordering
}
