% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.r
\name{ordering.maximise}
\alias{ordering.maximise}
\title{Find a good ordering in the sense that some function is
locally maximised.}
\usage{
ordering.maximise(ordering, fn)
}
\arguments{
\item{ordering}{The permutation (ordering) to start from.}

\item{fn}{A function to maximise.}
}
\description{
Find a good ordering in the sense that some function is
locally maximised.
}
