% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.r
\name{ordering.metropolis.hastings}
\alias{ordering.metropolis.hastings}
\title{Metropolis-Hastings on orderings.}
\usage{
ordering.metropolis.hastings(ordering, log.likelihood,
  proposal.fn = ordering.random.move, iterations = 1000, thin = 1)
}
\arguments{
\item{ordering}{Initial ordering}

\item{log.likelihood}{Log likelihood function}

\item{proposal.fn}{Proposal function}

\item{iterations}{Number of iterations}

\item{thin}{Thinning parameter}
}
\description{
Metropolis-Hastings on orderings.
}
