% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.r
\name{ordering.move}
\alias{ordering.move}
\title{Move one item in an ordering and shift the other items.}
\usage{
ordering.move(ordering, from, to)
}
\arguments{
\item{ordering}{The ordering.}

\item{from}{The index of the item to move.}

\item{to}{The index it should be moved to.}
}
\description{
Move one item in an ordering and shift the other items.
}
