% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.r
\name{ordering.random.block.move}
\alias{ordering.random.block.move}
\title{Randomly move a block in an ordering to another location}
\usage{
ordering.random.block.move(ordering, max.width = 4)
}
\arguments{
\item{ordering}{The ordering}

\item{max.width}{The maximum width of the block}
}
\description{
Randomly move a block in an ordering to another location
}

