% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.r
\name{ordering.random.move}
\alias{ordering.random.move}
\title{Randomly move one item in an ordering to another location}
\usage{
ordering.random.move(ordering)
}
\arguments{
\item{ordering}{The ordering.}
}
\description{
Randomly move one item in an ordering to another location
}

