% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.r
\name{ordering.test.score}
\alias{ordering.test.score}
\title{Test ordering score: sum every time consecutive items are in
order.}
\usage{
ordering.test.score(ordering)
}
\arguments{
\item{ordering}{The permutation (ordering) to test.}
}
\description{
Test ordering score: sum every time consecutive items are in
order.
}
