% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.r, R/model.r
\name{orderings.plot}
\alias{orderings.plot}
\alias{orderings.plot}
\title{Plot likelihoods of orderings against elapsed times taken
to generate them}
\usage{
orderings.plot(dl)

orderings.plot(dl)
}
\arguments{
\item{dl}{The DeLorean object}

\item{dl}{The DeLorean object}
}
\description{
Plot likelihoods of orderings against elapsed times taken
to generate them

Plot likelihoods of orderings against elapsed times taken
to generate them
}
