% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.r
\name{partition.de.lorean}
\alias{partition.de.lorean}
\title{Partition de.lorean object by cells}
\usage{
partition.de.lorean(dl, pieces = 2)
}
\arguments{
\item{dl}{de.lorean object}

\item{pieces}{How many pieces to partition into}
}
\description{
Partition de.lorean object by cells
}
