% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.r
\name{pca.expr}
\alias{pca.expr}
\title{Perform a PCA on the expression data. Calculate which direction in PC1-PC2 space
correlates best with capture time.}
\usage{
pca.expr(dl)
}
\arguments{
\item{dl}{The de.lorean object.}
}
\description{
Perform a PCA on the expression data. Calculate which direction in PC1-PC2 space
correlates best with capture time.
}

