% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothness.r
\name{permute.df}
\alias{permute.df}
\title{Permute a data frame, x. If group.col is given it should name an ordered
factor that the order of the permutation should respect.}
\usage{
permute.df(.df, group.col = NULL)
}
\arguments{
\item{.df}{Data frame}

\item{group.col}{Name of an ordered factor that the permutation
should respect.}
}
\description{
Permute a data frame, x. If group.col is given it should name an ordered
factor that the order of the permutation should respect.
}
