% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothness.r
\name{permuted.roughness}
\alias{permuted.roughness}
\title{Permute cells and test roughness of expression.}
\usage{
permuted.roughness(dl, expr.held.out = dl$expr.held.out)
}
\arguments{
\item{dl}{de.lorean object}

\item{expr.held.out}{The expression matrix of the held out genes}
}
\description{
Permute cells and test roughness of expression.
}
