% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.add.expr}
\alias{plot.add.expr}
\title{Add expression data to a plot}
\usage{
\method{plot}{add.expr}(gp, .data = NULL)
}
\arguments{
\item{gp}{Plot object}

\item{.data}{Expression data to add}
}
\description{
Add expression data to a plot
}
