% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{plot.add.mean.and.variance}
\alias{plot.add.mean.and.variance}
\title{Add posterior representation to a plot.}
\usage{
\method{plot}{add.mean.and.variance}(gp, .data = NULL, color = "black",
  line.alpha = 0.3, ribbon.alpha = 0.1)
}
\arguments{
\item{gp}{Plot object}

\item{.data}{Data frame containing variables to plot (mean, var)
phi, predictedvar)}

\item{color}{Color to use}

\item{line.alpha}{Alpha to use for mean line}

\item{ribbon.alpha}{Alpha to use for variance ribbon}
}
\description{
Add posterior representation to a plot.
}
