% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.de.lorean}
\alias{plot.de.lorean}
\title{Various DeLorean object plots}
\usage{
\method{plot}{de.lorean}(x, type = "profiles", ...)
}
\arguments{
\item{x}{de.lorean object}

\item{type}{Type of plot:
\itemize{
  \item 'expr.data': The expression data plotted by capture time.
    See \code{\link{expr.data.plot}}.
  \item 'Rhat': \eqn{hat{R}} convergence statistics
    See \code{\link{Rhat.plot}}.
  \item 'pseudotime': Pseudotimes in best posterior sample
    See \code{\link{pseudotime.plot}}.
  \item 'profiles': Gene expression profiles for best posterior sample
    See \code{\link{profiles.plot}}.
  \item 'tau.offsets': Offsets of pseudotimes to assess the prior
    See \code{\link{tau.offsets.plot}}.
  \item 'marg.like': Plot the posterior of the marginal likelihoods
    for individual genes.
    See \code{\link{marg.like.plot}}.
  \item 'roughnesses': Roughnesses of the pseudotime posterior
    See \code{\link{roughnesses.plot}}.
  \item 'init.vs.pseudotimes': Plot the initialisations used against the pseudotimes estimated
    See \code{\link{init.orderings.vs.pseudotimes.plot}}.
}}

\item{...}{Extra arguments to plot function}
}
\description{
Various DeLorean object plots
}

