% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/held-out.r
\name{plot.held.out.posterior}
\alias{plot.held.out.posterior}
\title{Plot the posterior of held out genes}
\usage{
\method{plot}{held.out.posterior}(dl, posterior, facets = ~gene)
}
\arguments{
\item{dl}{de.lorean object}

\item{posterior}{The posterior of some held out genes}

\item{facets}{Variables to wrap facets on}
}
\description{
Plot the posterior of held out genes
}
