% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_years.R
\name{plot_photos}
\alias{plot_photos}
\title{Plot available years for a given area}
\usage{
plot_photos(x)
}
\arguments{
\item{x}{Object of class \code{sf} or \code{sfc}}
}
\value{
data.frame
}
\description{
Plot available years for a given area
}
\examples{
\dontrun{
library(happign)
x <- get_apicarto_cadastre("29158")
photos <- plot_photos(x)
}

}
