% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{post.gene.psi.plot}
\alias{post.gene.psi.plot}
\title{Boxplot posterior of psi, the temporal variation parameter for each gene.}
\usage{
post.gene.psi.plot(dl)
}
\arguments{
\item{dl}{de.lorean object}
}
\description{
Boxplot posterior of psi, the temporal variation parameter for each gene.
}

