% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{post.omega.plot}
\alias{post.omega.plot}
\title{Plot posterior of omega, the noise parameter for all genes.}
\usage{
post.omega.plot(dl)
}
\arguments{
\item{dl}{de.lorean object}
}
\description{
Plot posterior of omega, the noise parameter for all genes.
}

