% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{post.psi.plot}
\alias{post.psi.plot}
\title{Plot posterior of psi, the temporal variation parameter for all genes.}
\usage{
post.psi.plot(dl)
}
\arguments{
\item{dl}{de.lorean object}
}
\description{
Plot posterior of psi, the temporal variation parameter for all genes.
}
