% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{prepare.for.stan}
\alias{prepare.for.stan}
\title{Prepare for Stan}
\usage{
prepare.for.stan(dl, num.test = 101, num.inducing = 30, period = 0,
  hold.out = 0)
}
\arguments{
\item{dl}{de.lorean object}

\item{num.test}{Number of test points to consider}

\item{num.inducing}{Number of inducing points}

\item{period}{Period of expression patterns}

\item{hold.out}{Number genes to hold out for generalisation tests}
}
\description{
Prepare for Stan
}

