% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{print.de.lorean}
\alias{print.de.lorean}
\title{Print details of DeLorean object}
\usage{
\method{print}{de.lorean}(x, ...)
}
\arguments{
\item{x}{de.lorean object}

\item{...}{Extra arguments}
}
\description{
Print details of DeLorean object
}
