% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{process.posterior}
\alias{process.posterior}
\title{Process the posterior, that is extract and reformat the samples from
Stan. We also determine which sample has the highest likelihood, this
is labelled as the 'best' sample.}
\usage{
process.posterior(dl)
}
\arguments{
\item{dl}{de.lorean object}
}
\description{
Process the posterior, that is extract and reformat the samples from
Stan. We also determine which sample has the highest likelihood, this
is labelled as the 'best' sample.
}

