% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{profiles.plot}
\alias{profiles.plot}
\title{Plot best sample predicted expression.}
\usage{
profiles.plot(dl, genes = dl$genes.high.psi, profile.color = "black",
  add.data = T, sample.iter = dl$best.sample, ignore.cell.sizes = FALSE,
  ...)
}
\arguments{
\item{dl}{de.lorean object}

\item{genes}{Genes to plot (defaults to genes.high.psi)}

\item{profile.color}{Colour for the profile}

\item{add.data}{Add actual expression data to plot}

\item{sample.iter}{Which sample to plot}

\item{ignore.cell.sizes}{Ignore cell sizes if the model has estimated them}

\item{...}{Extra arguments}
}
\description{
Plot best sample predicted expression.
}
