% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{pseudotime.plot}
\alias{pseudotime.plot}
\title{Plot pseudotime (tau) against observed capture time.}
\usage{
pseudotime.plot(dl, sample.iter = dl$best.sample)
}
\arguments{
\item{dl}{de.lorean object}

\item{sample.iter}{Which sample to take pseudotimes from}
}
\description{
Plot pseudotime (tau) against observed capture time.
}
