% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{pseudotimes.from.orderings}
\alias{pseudotimes.from.orderings}
\title{Convert best orderings into initialisations}
\usage{
pseudotimes.from.orderings(dl, num.to.keep = default.num.cores())
}
\arguments{
\item{dl}{The DeLorean object}

\item{num.to.keep}{The number to keep (defaults to default.num.cores())}
}
\description{
Convert best orderings into initialisations
}
