% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{pseudotimes.pair.plot}
\alias{pseudotimes.pair.plot}
\title{Plot two sets of pseudotimes against each other.}
\usage{
pseudotimes.pair.plot(dl, fits = NULL)
}
\arguments{
\item{dl}{The DeLorean object}

\item{fits}{Fit indexes}
}
\description{
Plot two sets of pseudotimes against each other.
}
