% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{psi.post.plot}
\alias{psi.post.plot}
\title{Plot the posterior of the temporal variation and show the prior}
\usage{
psi.post.plot(dl)
}
\arguments{
\item{dl}{The DeLorean object}
}
\description{
Plot the posterior of the temporal variation and show the prior
}
