% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{report.file}
\alias{report.file}
\title{The filename of the R markdown report.}
\usage{
report.file(report.name)
}
\arguments{
\item{report.name}{The report name}
}
\description{
The filename of the R markdown report.
}
