% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{Rhat.plot}
\alias{Rhat.plot}
\title{Plot the Rhat convergence statistics. \code{\link{examine.convergence}}
must be called before this plot can be made.}
\usage{
Rhat.plot(dl)
}
\arguments{
\item{dl}{de.lorean object}
}
\description{
Plot the Rhat convergence statistics. \code{\link{examine.convergence}}
must be called before this plot can be made.
}

