% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothness.r
\name{roughness.of.permutations}
\alias{roughness.of.permutations}
\title{Apply permutation based roughness test to held out genes}
\usage{
roughness.of.permutations(dl, expr.held.out = dl$expr.held.out,
  num.perms = 1000)
}
\arguments{
\item{dl}{de.lorean object}

\item{expr.held.out}{The expression matrix including the held out genes}

\item{num.perms}{Number of permutations to test}
}
\description{
Apply permutation based roughness test to held out genes
}
