% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothness.r
\name{roughness.of.sample}
\alias{roughness.of.sample}
\title{Calculate the roughness of the held out genes given the sample.}
\usage{
roughness.of.sample(dl, expr.held.out = dl$expr.held.out,
  sample.iter = dl$best.sample)
}
\arguments{
\item{dl}{de.lorean object}

\item{expr.held.out}{The expression matrix including the held out genes}

\item{sample.iter}{Which sample to use}
}
\description{
Calculate the roughness of the held out genes given the sample.
}
