% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothness.r
\name{roughness.test}
\alias{roughness.test}
\title{Calculate roughnesses under fit samples and also under random
permutations}
\usage{
roughness.test(dl, expr.held.out = dl$expr.held.out, num.perms = 1000)
}
\arguments{
\item{dl}{de.lorean object}

\item{expr.held.out}{The expression matrix including the held out genes}

\item{num.perms}{Number of permutations to test}
}
\description{
Calculate roughnesses under fit samples and also under random
permutations
}

