% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run Data Update Inside a Data Package Structure}
\usage{
run(
  update_processor,
  download_processor,
  data_processor,
  metadata_processor,
  data_dir = "data-raw"
)
}
\arguments{
\item{update_processor}{function that returns a POSIXct date
with timezone UTC to represent the date of the last time the publisher
updated the data.}

\item{download_processor}{function to obtain data from a file download,
database or API.}

\item{data_processor}{function to turn downloaded artefact into long format
csv}

\item{data_dir}{character relative data path reference.}

\item{meta_data_processor}{function to turn downloaded artefact meta data.
csv}
}
\description{
This wrapper runs all templated functions of a data package in standardized
fashion. This way bootstrapped data packages return standard output which is
compliant with opentimeseries. In order for this function to work properly
dataset specific functions need to be written by the enduser.
}
