% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_photos.R
\name{select_photos}
\alias{select_photos}
\title{Interactively select aerial photos to download}
\usage{
select_photos(photos, outdir = NULL)
}
\arguments{
\item{photos}{An \code{sf} object or \code{data.frame} produced by \code{\link[=find_photos]{find_photos()}}.}

\item{outdir}{\code{character} or \code{NULL}; Directory where photos will be saved.
Defaults to \code{NULL}, in which case the package user overwrite directory
(\code{\link[tools:userdir]{tools::R_user_dir()}}) is used.}
}
\value{
Invisibly returns a subset of \code{photos} corresponding to the user
selection. Returns \code{NULL} invisibly if the user quits.
}
\description{
Provides a simple console-based interactive workflow to help users select
a subset of photos returned by \code{\link[=find_photos]{find_photos()}}. The user chooses: a \strong{year},
a \strong{color type} and the number of photos to keep
}
\details{
This function does \strong{not} download files. It only returns the selected
rows (typically to be passed to \code{\link[=get_photos]{get_photos()}}).

Internal color codes are mapped as follows:
\itemize{
\item \code{"P"} -> \code{"PHOTO"} (panchromatic / black & white)
\item \code{"C"} -> \code{"RGB"} (color)
\item \verb{"IR"`` -> }"IRC"` (infrared)
\item \verb{"IRC"`` -> }"IRC"` (infrared color)
}
}
\examples{
\dontrun{
x <- sf::read_sf("my_area.gpkg")
photos <- find_photos(x)

sel <- select_photos(photos)

# Then download:
get_photos(sel$url, outdir = tempdir(), mode = "raw")
}

}
