% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.r
\name{seriation.find.orderings}
\alias{seriation.find.orderings}
\title{Use seriation package to find good orderings}
\usage{
seriation.find.orderings(dl, .methods = c("TSP", "R2E", "HC", "GW", "OLO"),
  scaled = c("scaled", "unscaled"), dim.red = c("none", "pca", "kfa", "ica",
  "mds"), dims = geom.series(base = 2, max = min(8, nrow(dl$expr) - 1)),
  num.cores = default.num.cores(), num.tau.to.keep = default.num.cores())
}
\arguments{
\item{dl}{de.lorean object}

\item{.methods}{The seriation methods to apply}

\item{scaled}{Whether to use the scaled and/or unscaled expression data}

\item{dim.red}{Dimension reduction methods to apply}

\item{dims}{Number of dimensions to reduce to}

\item{num.cores}{Number of cores to use in parallel}

\item{num.tau.to.keep}{How many initialisations to keep}
}
\description{
Use seriation package to find good orderings
}

