% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_common.R
\name{set_delorean_env_vars}
\alias{set_delorean_env_vars}
\title{Set Default deloRean Environment Variables}
\usage{
set_delorean_env_vars()
}
\description{
Set environment variables for timezone, author, author email and SSH
key location
}
