% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{tau.offsets.plot}
\alias{tau.offsets.plot}
\title{Plot the tau offsets, that is how much the pseudotimes (tau) differ
from their prior means over the full posterior.}
\usage{
tau.offsets.plot(dl, rug.alpha = 0.3)
}
\arguments{
\item{dl}{de.lorean object}

\item{rug.alpha}{Alpha parameter for rug geom}
}
\description{
Plot the tau offsets, that is how much the pseudotimes (tau) differ
from their prior means over the full posterior.
}
