% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{test.mh}
\alias{test.mh}
\title{Test ordering Metropolis-Hastings sampler.}
\usage{
test.mh(dl, psi = mean(dl$gene.map$psi.hat),
  omega = mean(dl$gene.map$omega.hat),
  num.cores = getOption("DL.num.cores", max(parallel::detectCores() - 1,
  1)), iterations = 1000, thin = 15)
}
\arguments{
\item{dl}{de.lorean object}

\item{psi}{Temporal variation}

\item{omega}{Noise}

\item{num.cores}{Number of cores to run on.
Defaults to getOption("DL.num.cores", max(parallel::detectCores()-1, 1))}

\item{iterations}{Number of iterations}

\item{thin}{Thin the samples}
}
\description{
Test ordering Metropolis-Hastings sampler.
}
