% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.r
\name{test.robustness.de.lorean}
\alias{test.robustness.de.lorean}
\title{Test robustness of pseudotime estimation on subsets of de.lorean
object}
\usage{
test.robustness.de.lorean(dl, pieces = 2)
}
\arguments{
\item{dl}{de.lorean object}

\item{pieces}{How many pieces to partition into}
}
\description{
Test robustness of pseudotime estimation on subsets of de.lorean
object
}
