% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_meta.R
\name{traverse_hierarchy}
\alias{traverse_hierarchy}
\title{Traverse Hierarchy to Get All Paths}
\usage{
traverse_hierarchy(hierarchy, current_path = character())
}
\arguments{
\item{hierarchy}{The hierarchy object (nested list).}

\item{current_path}{Current path being built (used in recursion).}
}
\value{
List of character vectors, each representing a path to a leaf.
}
\description{
Recursively traverses the hierarchy object and returns all
leaf paths as character vectors.
}
\keyword{internal}
