% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_seal.R
\name{version_seal}
\alias{version_seal}
\title{Finalize Version of a Set of Time Series}
\usage{
version_seal(v_obj)
}
\description{
This function commits a version to an archive and tags it accordingly with a
date. The opentimeseries package is designed to store versions as high as daily frequency.
higher frequency require workarounds
}
