% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{windram.expr}
\alias{windram.expr}
\alias{windram.cell.meta}
\alias{windram.gene.meta}
\title{Windram et al. investigated the defense response in Arabidopsis
thaliana to the necrotrophic fungal pathogen Botrytis cinerea.
They collected data at 24 time points in two conditions for
30336 genes.}
\format{There are three objects in this data:
\itemize{
  \item windram.expr A matrix of log expression values with
    no missing data. Rows are named by genes and columns are
    named by cells/samples.
  \item windram.gene.meta A data frame containing meta-data
    about the genes.
  \item windram.cell.meta A data frame containing meta-data
    about the cells
}}
\source{
\url{http://www.plantcell.org/content/24/9/3530.long}
}
\usage{
data(WindramDeLorean)
}
\description{
Windram et al. investigated the defense response in Arabidopsis
thaliana to the necrotrophic fungal pathogen Botrytis cinerea.
They collected data at 24 time points in two conditions for
30336 genes.
}
\keyword{datasets}
