% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-size.r
\name{adjust.by.cell.sizes}
\alias{adjust.by.cell.sizes}
\title{Adjust the expression by the estimated cell sizes.}
\usage{
adjust.by.cell.sizes(dl)
}
\arguments{
\item{dl}{de.lorean object.}
}
\description{
Adjust the expression by the estimated cell sizes.
}
