% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{branching.post.plot}
\alias{branching.post.plot}
\title{Plot the posterior of the branching model}
\usage{
branching.post.plot(dl, post, facets = ~gene)
}
\arguments{
\item{dl}{The DeLorean object}

\item{post}{The posterior (output of branching.genes.post)}

\item{facets}{Facets to wrap plots (passed to facet_wrap())}
}
\description{
Plot the posterior of the branching model
}
