% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{calc.inducing.pseudotimes}
\alias{calc.inducing.pseudotimes}
\title{Calculate inducing pseudotimes for sparse approximation}
\usage{
calc.inducing.pseudotimes(dl, num.inducing, period = 0)
}
\arguments{
\item{dl}{de.lorean object}

\item{num.inducing}{Number of inducing points}

\item{period}{Period of expression patterns}
}
\description{
Calculate inducing pseudotimes for sparse approximation
}

