% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{calc.inducing.pseudotimes}
\alias{calc.inducing.pseudotimes}
\title{Calculate inducing pseudotimes for sparse approximation}
\usage{
calc.inducing.pseudotimes(dl, num.inducing, period = 0, num.sd.border = 7)
}
\arguments{
\item{dl}{de.lorean object}

\item{num.inducing}{Number of inducing points}

\item{period}{Period of expression patterns}

\item{num.sd.border}{The size of the border of the inducing inputs
around the capture times in units of number of standard
deviations}
}
\description{
Calculate inducing pseudotimes for sparse approximation
}

