% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_common.R
\name{create_vintages}
\alias{create_vintages}
\title{Create Time Series Versions Data Tables (Vintages)}
\usage{
create_vintages(
  release_date,
  tsl,
  time_chunk = "23:59:59",
  tz = Sys.getenv("DELOREAN_TZ")
)
}
\arguments{
\item{release_date}{character date vector of release dates, will be
converted to POSIXct.}

\item{tsl}{list named list of time series.}

\item{time_chunk}{character time formatted as HH:MM:SS.}
}
\description{
This functions uses a vector of release dates and a list of time series to
create a nested data tables that contains all time series and their release dates.
}
